refactor: extract BuildingHelpers.cs — shared utilities from BuildingTools
